Domain-part update.
Changed the logic behind the map-pool selection feature.

Instead of it fetching contents of a folder in the WebPanel-folder, you now have to upload the map-pools yourself. 

This might be a bit annoying over time, but allows for simplification of setting up the overlay itself, completely negating the need for XAMPP/Apache, allowing the overlay to be run fully offline, without any extra software needed outside NodeJS for the relay-server.